Move test from shared sources
